# docs/api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: CBSC Strategy Management API
  description: |
    CBSC量化交易策略管理系統API文檔

    此API提供了完整的策略管理功能，包括：
    - 策略的創建、讀取、更新、刪除
    - 策略執行和監控
    - 實時市場數據獲取
    - 回測功能
    - 用戶管理

    # 認證
    API使用JWT Bearer Token進行認證。請在Authorization header中包含有效的token。

    # 錯誤處理
    API使用標準HTTP狀態碼。錯誤響應格式如下：
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "錯誤描述",
        "details": {}
      },
      "timestamp": "2025-01-19T10:00:00Z"
    }
    ```
  version: 2.0.0
  contact:
    name: CBSC API Support
    email: api-support@cbsc.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cbsc.example.com/v2
    description: Production server
  - url: https://staging-api.cbsc.example.com/v2
    description: Staging server
  - url: http://localhost:3004/v2
    description: Development server
security:
  - bearerAuth: []
...
paths:
...
  /strategies:
    get:
      tags:
        - Strategies
      summary: 獲取策略列表
      description: 分頁獲取用戶的策略列表，支持篩選和排序
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 頁碼，從1開始
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          description: 每頁數量
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          description: 策略狀態篩選
          schema:
            type: string
            enum: [active, inactive, archived, testing]
        - name: category
          in: query
          description: 策略分類
          schema:
            type: string
            enum: [momentum, mean_reversion, arbitrage, market_making]
        - name: sort_by
          in: query
          description: 排序字段
          schema:
            type: string
            enum: [created_at, updated_at, name, performance]
            default: created_at
        - name: sort_order
          in: query
          description: 排序順序
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 成功獲取策略列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      strategies:
                        type: array
                        items:
                          $ref: '#/components/schemas/Strategy'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Strategies
      summary: 創建新策略
      description: 創建一個新的交易策略
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyCreate'
      responses:
        '201':
          description: 策略創建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Strategy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /strategies/{strategy_id}:
    get:
      tags:
        - Strategies
      summary: 獲取策略詳情
      description: 根據策略ID獲取策略的詳細信息
      security:
        - bearerAuth: []
      parameters:
        - name: strategy_id
          in: path
          required: true
          description: 策略ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功獲取策略詳情
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StrategyDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Strategies
      summary: 更新策略
      description: 更新現有策略的信息
      security:
        - bearerAuth: []
      parameters:
        - name: strategy_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyUpdate'
      responses:
        '200':
          description: 策略更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Strategy'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Strategies
      summary: 刪除策略
      description: 軟刪除策略（標記為已刪除）
      security:
        - bearerAuth: []
      parameters:
        - name: strategy_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 策略刪除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Strategy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [momentum, mean_reversion, arbitrage, market_making]
        status:
          type: string
          enum: [active, inactive, archived, testing]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        performance:
          type: object
          properties:
            total_return:
              type: number
              format: float
            sharpe_ratio:
              type: number
              format: float
            max_drawdown:
              type: number
              format: float
    StrategyCreate:
      type: object
      required:
        - name
        - category
        - config
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        category:
          type: string
          enum: [momentum, mean_reversion, arbitrage, market_making]
        config:
          type: object
          description: 策略配置參數
        tags:
          type: array
          items:
            type: string
    StrategyDetail:
      allOf:
        - $ref: '#/components/schemas/Strategy'
        - type: object
          properties:
            config:
              type: object
              description: 策略配置
            execution_history:
              type: array
              items:
                $ref: '#/components/schemas/ExecutionRecord'
...
    ExecutionRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [running, paused, stopped]
        pnl:
          type: number
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未授權或token無效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...